// v3/internal/runtime/desktop/@wailsio/runtime/src/protocol.generated.ts
// Code generated by go generate; DO NOT EDIT.

export const objectNames = Object.freeze({
    Call: 0,
    Clipboard: 1,
    Application: 2,
    Events: 3,
    ContextMenu: 4,
    Dialog: 5,
    Window: 6,
    Screens: 7,
    System: 8,
    Browser: 9,
    CancelCall: 10,
    IOS: 11,
    Android: 12,
    EmbeddedWebView: 13,
});

export const embeddedWebViewMethods = Object.freeze({
    Create: 0,
    Destroy: 1,
    SetBounds: 2,
    SetVisible: 3,
    SetZIndex: 4,
    SetExclusions: 5,
    LoadURL: 6,
    GetURL: 7,
    GetTitle: 8,
    IsLoading: 9,
    Stop: 10,
    Reload: 11,
    ReloadIgnoringCache: 12,
    CanGoBack: 13,
    CanGoForward: 14,
    GoBack: 15,
    GoForward: 16,
    ExecuteJavaScript: 17,
    OpenDevTools: 18,
    SetZoomFactor: 19,
    GetZoomFactor: 20,
    Focus: 21,
    IsFocused: 22,
});